+++ CHANGES.md
@@ -4,3 +4,6 @@ Change Log
 ## changes in 0.2-SNAPSHOT
 
 * changed asImage to toDataURL
+
+* added ratio to def asCanvas(ratio: Double = 1)
+* added Outline ---> not yet finished need more info